+++ Editor/GUI/CustomEditor/EditorOnly/FaceTuneAssistantEditor.cs
@@ -1,3 +1,6 @@
+using aoyon.facetune.importer;
+using UnityEditor.Animations;
+
 namespace aoyon.facetune.gui
 {
     internal enum HandGesturePatternType
@@ -7,7 +10,6 @@ namespace aoyon.facetune.gui
         BasicRight,
         BasicLeft,
         Blending,
-        FaceMorphFirst,
         HandSign
     }
 
@@ -38,18 +40,16 @@ namespace aoyon.facetune.gui
         internal const string BasicRightPatternGuid = "c259edc6efd4aaa4bba3b1636557cc3b";
         internal const string BasicLeftPatternGuid = "376099cca4d264b4fbfbeeb7901dc770";
         internal const string BlendingPatternGuid = "9eb5bf9eeb8dc81488fb9453d21f3510";
-        internal const string FaceMorphFirstPatternGuid = "618bf06062904004f99355468c34ac7c";
         internal const string HandSignPatternGuid = "e7a261d8cf051454ea0c41e427463276";
 
         private static readonly Dictionary<HandGesturePatternType, PatternInfo> _handGesturePatternDetails = new()
         {
-            { HandGesturePatternType.LeftOnly, new PatternInfo("左手のみのパターンです。", LeftOnlyPatternGuid) },
-            { HandGesturePatternType.RightOnly, new PatternInfo("右手のみのパターンです。", RightOnlyPatternGuid) },
-            { HandGesturePatternType.BasicRight, new PatternInfo("右手が優先される基本的なパターンです。", BasicRightPatternGuid) },
-            { HandGesturePatternType.BasicLeft, new PatternInfo("左手が優先される基本的なパターンです。", BasicLeftPatternGuid) },
-            { HandGesturePatternType.Blending, new PatternInfo("片手ごとのアニメーションがブレンドさせるパターンです。目と口の制御をそれぞれの手に割り当てる際などに便利です。", BlendingPatternGuid) },
-            { HandGesturePatternType.FaceMorphFirst, new PatternInfo("右手と左手に優先度を付けず、最初に実行したジェスチャーを優先させるパターンです。", FaceMorphFirstPatternGuid) },
-            { HandGesturePatternType.HandSign, new PatternInfo("左手と右手の組み合わせで最大64通りのジェスチャーを作成できるパターンです。", HandSignPatternGuid) },
+            { HandGesturePatternType.LeftOnly, new PatternInfo("左手のみのパターンです。(8通り)", LeftOnlyPatternGuid) },
+            { HandGesturePatternType.RightOnly, new PatternInfo("右手のみのパターンです。(8通り)", RightOnlyPatternGuid) },
+            { HandGesturePatternType.BasicRight, new PatternInfo("右手が優先される基本的なパターンです。(16通り)", BasicRightPatternGuid) },
+            { HandGesturePatternType.BasicLeft, new PatternInfo("左手が優先される基本的なパターンです。(16通り)", BasicLeftPatternGuid) },
+            { HandGesturePatternType.HandSign, new PatternInfo("左手と右手の組み合わせで最大64通りのジェスチャーを作成できるパターンです。(64通り)", HandSignPatternGuid) },
+            { HandGesturePatternType.Blending, new PatternInfo("片手ごとのアニメーションがブレンドさせるパターンです。目と口の制御をそれぞれの手に割り当てる際などに便利です。(64通り)", BlendingPatternGuid) },
         };
 
         private static readonly Dictionary<OtherPatternType, PatternInfo> _otherPatternDetails = new()
@@ -61,6 +61,7 @@ namespace aoyon.facetune.gui
 
         private HandGesturePatternType _selectedHandGesturePattern = HandGesturePatternType.BasicRight;
         private OtherPatternType _selectedOtherPattern = OtherPatternType.ExclusiveMenu;
+        private AnimatorController? _selectedAnimatorController;
 
         private PresetComponent[] _presets = Array.Empty<PresetComponent>();
         private ConditionComponent[] _conditions = Array.Empty<ConditionComponent>();
@@ -73,7 +74,7 @@ namespace aoyon.facetune.gui
 
         public override void OnInspectorGUI()
         {
-            EditorGUILayout.LabelField("ExpressionコンポーネントはHierarhy上で下にあるほど優先度が高くなります。", EditorStyles.boldLabel);
+            EditorGUILayout.LabelField("ExpressionコンポーネントはHierarhyにおいて下にあるほど優先度が高くなります。", EditorStyles.boldLabel);
 
             // --- サジェスチョン表示エリアここから ---
             EditorGUILayout.Space();
@@ -108,7 +109,7 @@ namespace aoyon.facetune.gui
         internal void CreatePatternGUI()
         {
             EditorGUILayout.BeginVertical(GUI.skin.box);
-            EditorGUILayout.LabelField("サンプルパターンを追加", EditorStyles.boldLabel);
+            EditorGUILayout.LabelField("パターンを追加", EditorStyles.boldLabel);
 
             EditorGUILayout.Space();
 
@@ -139,12 +140,37 @@ namespace aoyon.facetune.gui
                 CreatePatternImpl(_otherPatternDetails[_selectedOtherPattern]);
             }
             EditorGUILayout.EndVertical();
+
+            EditorGUILayout.Space();
+
+            EditorGUILayout.BeginVertical(EditorStyles.helpBox);
+            EditorGUILayout.LabelField("Animator Controllerをインポート", EditorStyles.boldLabel);
+            _selectedAnimatorController = (AnimatorController)EditorGUILayout.ObjectField("選択中のAnimator Controller:", _selectedAnimatorController, typeof(AnimatorController), false);
+            if (GUILayout.Button("追加"))
+            {
+                ImportAnimatorController();
+            }
+            EditorGUILayout.EndVertical();
             EditorGUILayout.EndVertical();
         }
 
         private void CreatePatternImpl(PatternInfo patternInfo)
         {
             FTPrefabUtility.InstantiatePrefab(patternInfo.Guid, true, Component.gameObject);
+        }
+
+        private void ImportAnimatorController()
+        {
+            if (_selectedAnimatorController == null)
+            {
+                throw new Exception("Animator Controller is not selected");
+            }
+            if (!CustomEditorUtility.TryGetContext(Component.gameObject, out var context))
+            {
+                throw new Exception("Failed to get context");
+            }
+            var importer = new AnimatorControllerImporter(context, _selectedAnimatorController);
+            importer.Import(Component.gameObject);
         }
     }
 }