+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Japanese version](./CHANGELOG-jp.md)
+## [日本語](./CHANGELOG-jp.md)
 
 ## [Unreleased]
 ### Added
@@ -14,6 +14,23 @@
 ### Fixed
 
 ### Security
+
+## [0.1.0-beta.5] - 2025-08-05
+### Added
+- Added import function for Animator Controller to `FaceTune Assistant`.
+- Specified platform-dependent build behavior.
+  - Application to Animator is performed only for VRChat builds; other passes work on all platforms.
+
+### Changed
+- Changed the specification for importing Animator Controllers.
+  - Expressions are now generated only if blendshapes for facial expressions are included.
+  - If both facial animation and non-facial animation are included, `Animation Data` will be generated.
+  - The generated `Facial Data` now defaults to Manual Mode.
+- When generating patterns from `FaceTune Assistant`, Prefabs are now fully unpacked.
+
+### Removed
+- Removed the menu item `GameObject/FaceTune/Import from FX Layer`.
+- Removed the FaceMorphFirst pattern.
 
 ## [0.1.0-beta.4] - 2025-08-03
 ### Data for almost all FaceTune components will be lost.